Add make_perl_context script
This creates a new docker context, from which a new docker image
can be built, based on an official Perl image for the specified
perl version.

This image can be used in place of the official
`ledgersmb/ledgersmb-dev-lsmb` image for testing against
specific version of Perl.

The standard context provided in ./ledgersmb is used a a template
for the newly created context.